mat_mult.c
This is an example of implementing parallel matrix multiplication
using the Message Passing Interface (MPI). The goal of the program
is to distribute the workload of matrix multiplication across
multiple processes, allowing for faster computation of large
matrices. 

Steps:
● Matrix Initialization- A & B
● Matrix Broadcasting
The initialized matrices are broadcasted to all other processes
using the MPI_Bcast function. This ensures that each process has
access to the input matrices
● Parallel Computation:
The matrices are divided into segments, and each process is
assigned a portion of the computation.
The resulting matrix is computed by multiplying the
corresponding elements of the divided matrices.
● MPI_Gather for Result Collection
● Output Display
 The program displays the original matrices (A and B) and the final result matrix on the root process.